Customer Api done for now

+++ app/Http/Controllers/Api/FoodController.php
@@ -6,7 +6,9 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreFoodRequest;
 use App\Http\Requests\UpdateFoodRequest;
 use App\Models\Food;
-
+use App\Policies\FoodPolicy;
+use Illuminate\Database\Eloquent\Attributes\UsePolicy;
+#[UsePolicy(FoodPolicy::class)]
 class FoodController extends Controller
 {
     /**
@@ -31,46 +33,66 @@ class FoodController extends Controller
      */
     public function store(StoreFoodRequest $request)
     {
-        // Ez fogja meghívni a Policy-t
-        $this->authorize('create', Food::class);
+        if ($this->authorize('create', Food::class)) {
+            $food = Food::create($request->validated());
 
-        $food = Food::create($request->validated());
+            return response()->json([
+                'msg' => 'Created successfully',
+                'data' => $food,
+            ], 201);
+        }
 
-        return response()->json([
-            'msg' => 'Created successfully',
-            'data' => $food,
-        ], 201);
+        return response()->json(['msg' => 'Unauthorized'], 403);
+
     }
 
     /**
      * Display the specified resource.
      */
-    public function show(string $id)
+    public function show(string $foodName)
     {
-        $food = Food::findOrFail($id);
+        if ($food = Food::where('name', $foodName)->first()) {
+            return response()->json($food);
+        }
 
-        return response()->json($food);
+        return response()->json(['msg' => "$foodName not found"], 404);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(UpdateFoodRequest $request, string $id)
+    public function update(UpdateFoodRequest $request, string $foodName)
     {
-        $food = Food::findOrFail($id);
-        $food->update($request->all());
+        if ($this->authorize('update', Food::class)) {
+            if ($food = Food::where('name', $foodName)->first()) {
+                $food->update($request->all());
 
-        return response()->json(['msg' => "{$food->name} was updated"]);
+                return response()->json(['msg' => "{$food->name} was updated", 'data' => $food]);
+            }
+
+            return response()->json(['msg' => "$foodName not found"], 404);
+        }
+
+        return response()->json(['msg' => 'Unauthorized'], 403);
+
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(string $foodName)
     {
-        $food = Food::findOrFail($id);
-        $food->delete();
+        if ($this->authorize('delete', Food::class)) {
+            if ($food = Food::where('name', $foodName)->first()) {
+                $food->delete();
 
-        return response()->json(['msg' => "{$food->name} was deleted"]);
+                return response()->json(['msg' => "{$food->name} was deleted"]);
+            }
+
+            return response()->json(['msg' => "$foodName not found"], 404);
+        }
+
+        return response()->json(['msg' => 'Unauthorized'], 403);
+
     }
 }
